Spanify: Add trailing comma in initializer list.
Clang format offers a better formatting when a trailing comma is used,
because it puts each element on a separate line. Everything is aligned,
and the code is more readable.

This is particularly helpful when the original code not formatted by
clang-format and already on multiple lines. Not adding the trailing
comma would cause an unfortunate reformatting.

To allow for one-liner, omit this step for small initialization lists.

Bug: 381024732
Change-Id: I268d081f1989f3a1d8146e39693aa0437f4a0a7f
Fixed: 381024732
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6049960
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1389284}
NOKEYCHECK=True
GitOrigin-RevId: 379a4b6b39e793c795b917ecd6b32afa1fde5985