[blink_gc_plugin] Eliminate false alerts
The plugin was alerting that `TraceIfNeeded<std::unique_ptr>` is
illegally tracing a field. `TraceIfNeeded` is used for templates when
the type is not always known. `TraceIfNeeded<std::unique_ptr>` is the
intended usage and the plugin should not alert on it.

Bug: 40248881
Change-Id: Ief1937d9a649c2a7229149b8da6136aba9a4fa53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6187142
Reviewed-by: David Baron <dbaron@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410156}
NOKEYCHECK=True
GitOrigin-RevId: a52d7b7471f6b08ff912be1777279dfb164f305e