Spanification: Represent source nodes explicitly.
This change is purely refactorial and does not affect the behavior of
the spanification tool.

Parent patch:
https://chromium-review.googlesource.com/c/chromium/src/+/6190594

Declaring the sources of a rewrite used two different instructions was
part of the Node, which could be emitted multiple times from:
- An edge.
- A single Node.

The boolean was not consistent, and there was some custom logic to
mixe/deduplicate this data.

To make the logic simpler and easier to read. We move this as a "source"
serialized instruction.

Bug: 356643982
Change-Id: I277eb008f0c01c15fc092bf46a838ce6f6fa9088
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6191878
Reviewed-by: Ali Hijazi <ahijazi@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Auto-Submit: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Ali Hijazi <ahijazi@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411653}
NOKEYCHECK=True
GitOrigin-RevId: 0b096e879efeed55b466cf1b9ca6fbf43d893fee