Don't assume .out exists if we haven't compiled.
While handling the exception of not applying a patch I copied the
failure branch of compiling which attempts to copy the .out file to a
.fail file. However this .out only exists if we attempted to compile.

Instead we now just create the .fail file directly and output the
stderr/stdout like it would have contained.

We got an exception like the following:
Traceback (most recent call last):
  File "/usr/local/google/home/nuskos/chromium-team/chromium/src/tools/clang/spanify/evaluate_patches.py", line 274, in <module>
    shutil.copy(scratch_dir + f"/patch_{index}.out",
  File "/usr/lib/python3.12/shutil.py", line 435, in copy
    copyfile(src, dst, follow_symlinks=follow_symlinks)
  File "/usr/lib/python3.12/shutil.py", line 260, in copyfile
    with open(src, 'rb') as fsrc:

Also wrap errormsg in quotes so that if there is new lines in the
exception it doesn't break the CSV format in evalutions.csv

Finally also run `git restore .` because the failed patch often leaves a partially modified change and we don't commit it so it goes to the next branch.

Change-Id: I39b2773465329d4f37279c17836ca9d30f695c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6311634
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1426927}
NOKEYCHECK=True
GitOrigin-RevId: eebcb3142588f6cb5f1fcf0cd5dca158a905a566