Don't continue the evaluate_patches when tests fail
When we experience a failure in the rewrite script it doesn't make sense
to continue due to not having any actual patches to attempt.

When the current test was failing it would attempt to apply test patches
to the main chromium repo and than fail to git commit.

Change-Id: I1e19e7702700b4b7d62a51d46fea7a0f8c3a5f54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6326291
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Stephen Nusko <nuskos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1428841}
NOKEYCHECK=True
GitOrigin-RevId: 7e99c8f5840a692e79150618b57e4f38dc0d3795