Simple AST rewriter example
This is an implementation of an AST rewriter based on clang's AST
matchers, designed to be usable as a template for learning how to
make such rewriters/beginning new ones. The rewriter itself
replaces instances of `b ? "true" : "false"` with `base::ToString(b)`
in C++ code, and was used to make several contributions to
crbug.com/335797528.

Change-Id: I0602694ed1acb424cc3e5eb17a8e89837ded4350
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6322829
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1430219}
NOKEYCHECK=True
GitOrigin-RevId: 3b59be9469b8ec236a4a165b4647b801c41769e0