spanify: Fix a bug for rewriting x => !x.empty()
When rewriting an implicit conversion from `x` to a boolean,
we're rewriting it to `!x.empty()`. But the previous patch
https://crrev.com/c/6280526 was wrong about the position to
insert `!` in case of a macro, e.g. MACRO(x). It inserts `!`
into a macro definition despite that `!` should be inserted
before `x` on the call sites.

New test cases excersise the regression cases.

Bug: 397043083, 358306232
Change-Id: I4cdbc7255986e110204408eaaf724cea07797467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6332463
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Reviewed-by: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1431341}
NOKEYCHECK=True
GitOrigin-RevId: ce0396f5d376d98648f16f4c2fd453a8944b7105