Handle empty commits without erroring.
We sometimes generate a patch that applies to third_party (this is a
bug), but when apply_edits sees it they ignore it. This causes an empty
commit which then fails when we attempt to commit. Handle that case as a
failure and upload to track occurances.

Change-Id: I008c5ddd322a813a745f8bfba0ef232fbed72291
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6355824
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1432558}
NOKEYCHECK=True
GitOrigin-RevId: 6decbf9beaa89f7979517c4cc898efd3667481e8