Clarify -mbranch-protection to use bti+pac-ret
Guarded control stack support for clang was added back in January, which
changes -mbranch-protection=standard to also mark objects as supporting
GCS on applicable architectures.

This is undesirable for Chromium as GCS could "silently" be enabled when
GCS-capable hardware becomes available. GCS has also not been explicitly
enabled in components such as libunwind.

arm_control_flow_integrity has been adjusted to clarify that it is
bti/pac-ret that is being used. So when the time comes to enable gcs it
may done so explicitly.

Change-Id: Ic0e630d4bc8cf27cc8a2a92eaed0b13353cb8a66
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6074598
Reviewed-by: Sam Maier <smaier@chromium.org>
Reviewed-by: Alan Zhao <ayzhao@google.com>
Commit-Queue: Jonathan Wright <jonathan.wright@arm.com>
Cr-Commit-Position: refs/heads/main@{#1453172}
NOKEYCHECK=True
GitOrigin-RevId: 284941756678030f68a4dc811c719da94b4af114