Filter -fprofile-use when building the compilation database
The flag -fprofile-use is used for PGO and is not required by tools.
Clangd complains it cannot find the referenced profile file and
generates an error for each open .cc or .h file. So filter the flag when
building the compilation database.

Change-Id: I705f20a6d56f708ba6c8bc806947b1aa26e4fc48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6675444
Reviewed-by: Aldo Culquicondor <acondor@chromium.org>
Commit-Queue: Anand Ravi <anandrv@google.com>
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1479099}
NOKEYCHECK=True
GitOrigin-RevId: 416dfa95eb6b52053fb40fdc216143a6d706dbcb