Support sanitized android runtime on x86_64
This is part of an effort to support coverage guided fuzzers on android
(arm64, x86_64).

This change specifically focuses on the clang binaries for sanitizers on
android, which were missing to be able to build the fuzzers on
android_x64 which is a platform/architecture we want to test.

Arm64 is aleady supported.

Android i686 had support for ASan but not UBSan, which is also corrected
here.

R=aeubanks@google.com

Bug: 328559555
Change-Id: I793a9a213eb8c840451890f7db30bd69e039f804
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6772919
Auto-Submit: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Arthur Eubanks <aeubanks@google.com>
Commit-Queue: Arthur Eubanks <aeubanks@google.com>
Cr-Commit-Position: refs/heads/main@{#1496484}
NOKEYCHECK=True
GitOrigin-RevId: 27be8b77710f4405fdfeb4ee946fcabb0f6c92b2