Update the preprocessed source code used to train Clang
A change in Clang made Clang no longer expand macros for already
preprocessed files, and this caused Clang to reject the existing
file we use for PGO training as that file included `xmmintrin.h`. The
newly regenerated preprocessed file does not include that header.

The file name ends with `-3.ii` instead of `-2.ii` as `-2.ii` already
exists in the GCS bucket. I have no idea why that exists.

Also, update the C++ version to C++20 in order for this to work.

Bug: 436324009
Change-Id: Ib33559052ea04013166e5c85f7408611e83249fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6821902
Commit-Queue: Alan Zhao <ayzhao@google.com>
Reviewed-by: Arthur Eubanks <aeubanks@google.com>
Cr-Commit-Position: refs/heads/main@{#1497216}
NOKEYCHECK=True
GitOrigin-RevId: 6300c2f85ffc843435756f401507f246a7420bd0