Write stamp files to actual build directory
If the user passes in a custom build directory, we shouldn't
edit the stamp files for the default directory where chromium's
clang lives, because we haven't actually changed it. Make sure
the stamp files are relative to the directory we're building
in.

Bug: 439887312
Change-Id: Ic35f56cf00f42448b42731b5a444953c3d3bf394
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6868589
Auto-Submit: Devon Loehr <dloehr@google.com>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504485}
NOKEYCHECK=True
GitOrigin-RevId: 2911bc068d7856debf76619b21caeea7be2727cb