Prepare unsafe buffers plugin for syntax expansion.
In order to allow the plugin to operate in situations where the build
directory is not a direct descendant of the source tree root, more
information is required in the unsafe buffers paths file: the location
of the source root relative to this file. In many products, the unsafe
buffers paths file is located in the source root itself (e.g. PDFium,
Angle) but for Chromium, it is located in build/config/. While we could
hard-code such knowledge into the plugin, a better approach, and one
discussed when initially placing this file for Chromium, is to add
syntax to the file to express this relationship.

Because of the need for compatibility across compiler rolls, this
is a three-phase process:
1. Update the plugin to tolerate and ignore the new syntax.
2. Update the unsafe buffers paths file with the new syntax.
3. Update the plugin to process the new syntax as intended.

This is step 1 of the above.

Because the unsafe buffers paths file is designed to be mindlessly
simple to parse, we introduce a new symbol for such lines along the
lines of the existing +,-, and . lines: a $ followed by the relative
path from the unsafe buffers path file to the source root. For
example, Chromium will specify:

$../../

Logically, this can be considered something to be prepended to
the path in every +/- line listing paths, in order to make that
path accurate with respect to the directory that this file is in.

-- Also reserve and tolerate @-lines for a future experiment.

Bug: 439019034
Change-Id: I5c8c3bb62775e7a7ad998adafbb5692d790440da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6861996
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1504748}
NOKEYCHECK=True
GitOrigin-RevId: e956b96117a892eb81a437dbd2f7189529fbbfbc