"Fix" the spanify rewrite-multiple-platforms script
Currently this seems to be failing with something related to
get_tensorflow.py which can be worked around by not using the
ml-inliner-model (by setting it to an empty string).

Seems strange but works, there probably is a better way to support this
but don't currently have time to investigate and we don't care about
inlining decisions in this project.

Change-Id: I843e0add3ea45bffd059094d84f0a22f7e13b908
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7201523
Auto-Submit: Stephen Nusko <nuskos@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1550910}
NOKEYCHECK=True
GitOrigin-RevId: f551447fe0a1b76af2b32bbf4a942f847828a830